Sync db with blog
