+++ part13/blog-app/index.js
@@ -1,75 +1,20 @@
-require("dotenv").config();
-const { Sequelize, Model, DataTypes } = require("sequelize");
 const express = require("express");
 const app = express();
 app.use(express.json());
 
-const sequelize = new Sequelize(process.env.DATABASE_URL, {
-  dialectOptions: {
-    ssl: {
-      require: true,
-      rejectUnauthorized: false,
-    },
-  },
-});
-class Blog extends Model {}
+const { PORT } = require("./util/config");
+const { connectToDatabase } = require("./util/db");
 
-Blog.init(
-  {
-    id: {
-      type: DataTypes.INTEGER,
-      primaryKey: true,
-      autoIncrement: true,
-    },
-    author: {
-      type: DataTypes.TEXT,
-    },
-    url: {
-      type: DataTypes.TEXT,
-      allowNull: false,
-    },
-    title: {
-      type: DataTypes.TEXT,
-      allowNull: false,
-    },
-    likes: {
-      type: DataTypes.INTEGER,
-      defaultValue: 0,
-    },
-  },
-  {
-    sequelize,
-    underscored: true,
-    timestamps: false,
-    modelName: "blog",
-  }
-);
+const blogsRouter = require("./controllers/blogs");
+app.use(express.json());
 
-Blog.sync();
+app.use("/api/blogs", blogsRouter);
 
-app.get("/api/blogs", async (req, res) => {
-  const blogs = await Blog.findAll();
-  res.json(blogs);
-});
-
-app.post("/api/blogs", async (req, res) => {
-  const blog = await Blog.create(req.body);
-  res.json(blog);
-});
-
-
-app.delete("/api/blogs/:id", async (req, res) => {
-  const deleted = await Blog.destroy({
-    where: { id: req.params.id }
+const start = async () => {
+  await connectToDatabase();
+  app.listen(PORT, () => {
+    console.log(`Server running on port ${PORT}`);
   });
-  if (!deleted) {
-    return res.status(404).json({ error: "Blog not found" });
-  }
-  res.status(204).end();
-});
+};
 
-
-const PORT = process.env.PORT || 3001;
-app.listen(PORT, () => {
-  console.log(`Server running on port ${PORT}`);
-});
+start();
