+++ part13/blog-app/models/index.js
@@ -1,0 +1,7 @@
+const Blog = require("./blog");
+
+Blog.sync();
+
+module.exports = {
+  Blog,
+};
